Complete Windows multiprocess WAL port
