+++ .github/workflows/rust.yml
@@ -363,8 +363,14 @@ jobs:
           prefix-key: "v1-rust"
           cache-on-failure: true
       - uses: "./.github/shared/setup-sccache"
-      - name: Run Windows multiprocess regression
-        run: cargo test -p turso_core --features "fs experimental_win_iocp" subprocess_database_open_selects_multiprocess_shm_backend
+      - name: Run Windows concurrent simulator multiprocess regressions
+        run: cargo test -q -p turso_whopper --test regression_tests
+      - name: Run Windows concurrent simulator cross-platform regression
+        run: cargo test -q -p turso_whopper --test regression_tests_cross_platform
+      - name: Run Windows CLI multiprocess VFS regression
+        run: cargo test -q -p turso_cli input::tests::experimental_win_iocp_backend_is_available_for_path_databases -- --exact
+      - name: Run Windows core multiprocess regressions
+        run: cargo test -q -p turso_core --features "fs experimental_win_iocp" multiprocess_tests:: -- --nocapture
       - name: Run Windows multiprocess startup smoke
         run: cargo run -q -p turso_whopper -- --mode fast --multiprocess --connections-per-process 1 --processes 2 --max-steps 0
