+++ cli/input.rs
@@ -192,6 +192,18 @@ pub fn get_io(db_location: DbLocation, io_choice: &str) -> anyhow::Result<Arc<dy
     })
 }
 
+#[cfg(all(test, target_os = "windows", feature = "experimental_win_iocp"))]
+mod tests {
+    use super::{DbLocation, get_io};
+
+    #[test]
+    fn experimental_win_iocp_backend_is_available_for_path_databases() {
+        let io = get_io(DbLocation::Path, "experimental_win_iocp")
+            .expect("windows cli should construct the experimental_win_iocp backend");
+        drop(io);
+    }
+}
+
 pub struct ApplyWriter<'a> {
     target: &'a Arc<turso_core::Connection>,
     // accumulate raw bytes to support non-utf8 BLOB types
